Countdown and Time fragments coverted to activities
