+++ app/src/main/AndroidManifest.xml
@@ -18,6 +18,14 @@
             </intent-filter>
         </activity>
         <activity
+            android:name=".time.TimeActivity"
+            android:label="@string/app_name" />
+
+        <activity
+            android:name=".countdown.CountdownActivity"
+            android:label="@string/app_name" />
+
+        <activity
             android:name=".AboutActivity"
             android:label="@string/title_activity_about_activity"/>
     </application>
