+++ app/src/main/java/com/willrussell/hackathon_demo_timer_android/time/TimeActivity.java
@@ -1,16 +1,20 @@
-package com.willrussell.hackathon_demo_timer_android;
+package com.willrussell.hackathon_demo_timer_android.time;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.os.Handler;
-import android.os.Looper;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.widget.TextView;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-import com.willrussell.hackathon_demo_timer_android.time_threads.TimeThread;
+import com.willrussell.hackathon_demo_timer_android.R;
+import com.willrussell.hackathon_demo_timer_android.AboutActivity;
+import com.willrussell.hackathon_demo_timer_android.countdown.CountdownActivity;
+
+import java.text.SimpleDateFormat;
+import java.util.Date;
 
 public class TimeActivity extends AppCompatActivity {
 
@@ -23,11 +27,34 @@ public class TimeActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_time);
 
-        handler = new Handler(Looper.getMainLooper());
-        timeView = findViewById(R.id.time_remaining);
+        handler = new Handler();
+        timeView = findViewById(R.id.time);
 
         thread = new TimeThread();
         new Thread(thread).start();
+    }
+
+    class TimeThread implements Runnable {
+        private String time;
+
+        public TimeThread() {
+
+        }
+
+        @Override
+        public void run() {
+            while (!Thread.currentThread().isInterrupted()) {
+                time = getTime();
+                handler.post(new UiThread(time, timeView));
+            }
+        }
+
+        public String getTime() {
+            SimpleDateFormat formatter = new SimpleDateFormat("HH:mm:ss");
+            Date date = new Date();
+            String time = formatter.format(date);
+            return time;
+        }
     }
 
     @Override
