+++ app/src/main/res/layout/activity_countdown.xml
@@ -7,16 +7,6 @@
     android:gravity="center">
 
     <TextView
-        android:id="@+id/connected_status"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/disconnected"
-        android:textColor="@color/disconnected"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent" />
-
-    <TextView
         android:id="@+id/time_remaining"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
